@@ -1,11 +1,6 @@
-apply plugin: "java"
-
-sourceCompatibility = 1.6
+sourceCompatibility = 1.7
 [compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
 
 sourceSets.main.java.srcDirs = [ "src/" ]
 
-
-eclipse.project {
-    name = appName + "-core"
-}
+eclipse.project.name = appName + "-core"
