@@ -61,6 +61,11 @@ public class TutorialScreen implements Screen {
             }
 
             @Override
+            public boolean touchCancelled(int screenX, int screenY, int pointer, int button) {
+                return false;
+            }
+
+            @Override
             public boolean touchDragged(int screenX, int screenY, int pointer) {
                 return false;
             }
@@ -71,7 +76,7 @@ public class TutorialScreen implements Screen {
             }
 
             @Override
-            public boolean scrolled(int amount) {
+            public boolean scrolled(float amountX, float amountY) {
                 return false;
             }
         });
