@@ -11,11 +11,14 @@ public class HtmlLauncher extends GwtApplication {
 
         @Override
         public GwtApplicationConfiguration getConfig () {
-                return new GwtApplicationConfiguration(480, 320);
+                // Resizable application, uses available space in browser
+                return new GwtApplicationConfiguration(true);
+                // Fixed size application:
+                //return new GwtApplicationConfiguration(480, 320);
         }
 
         @Override
-        public ApplicationListener getApplicationListener () {
+        public ApplicationListener createApplicationListener () {
                 return new BulletProof(new DummyScoreBoard(), new DummyShareResolver());
         }
 }
