@@ -1,5 +1,7 @@
 package org.liberty.multi.bulletproof;
 
+import org.liberty.multi.bulletproof.resolver.DummyScoreBoard;
+import org.liberty.multi.bulletproof.resolver.DummyShareResolver;
 import org.robovm.apple.foundation.NSAutoreleasePool;
 import org.robovm.apple.uikit.UIApplication;
 
@@ -11,7 +13,7 @@ public class IOSLauncher extends IOSApplication.Delegate {
     @Override
     protected IOSApplication createApplication() {
         IOSApplicationConfiguration config = new IOSApplicationConfiguration();
-        return new IOSApplication(new BulletProof(), config);
+        return new IOSApplication(new BulletProof(new DummyScoreBoard(), new DummyShareResolver()), config);
     }
 
     public static void main(String[] argv) {
